Feature/loading profiling (#154)
* [Add] tracy profiling for model loading [Moved] Old engine code to Game play module

* [Add] Missing include

* [Removed] unused and deprecated files [Removed] unused includes [Removed] OldEngine target

* [Fix] include errors and broken references to OldEngine
