+++ engine/bindings/private/wren_bindings.cpp
@@ -11,7 +11,6 @@
 #include "ecs_module.hpp"
 #include "game/game_bindings.hpp"
 #include "input/input_bindings.hpp"
-#include "lifetime_component.hpp"
 #include "particle_module.hpp"
 #include "particles/particle_bindings.hpp"
 #include "pathfinding/pathfinding_bindings.hpp"
@@ -19,12 +18,10 @@
 #include "physics/physics_bindings.hpp"
 #include "physics_module.hpp"
 #include "renderer/animation_bindings.hpp"
+#include "systems/lifetime_component.hpp"
 #include "time_module.hpp"
-#include "utility/enum_bind.hpp"
 #include "utility/wren_entity.hpp"
 #include "wren_engine.hpp"
-
-#include <cstdint>
 
 namespace bindings
 {
