+++ engine/game/private/serialization.hpp
