+++ engine/game/public/scene/scene_loader.hpp
