+++ engine/inspector/private/editor.hpp
@@ -3,7 +3,6 @@
 #include "common.hpp"
 #include "ecs_module.hpp"
 #include "imgui_entt_entity_editor.hpp"
-#include <memory>
 
 class PerformanceTracker;
 class BloomSettings;
