+++ engine/inspector/private/inspector_module.cpp
@@ -4,9 +4,9 @@
 #include "gpu_scene.hpp"
 #include "graphics_context.hpp"
 #include "imgui_backend.hpp"
-#include "implot/implot.h"
-#include "menus/particle_editor.hpp"
+#include "implot.h"
 #include "magic_enum.hpp"
+#include "menus/particle_editor.hpp"
 #include "menus/performance_tracker.hpp"
 #include "renderer.hpp"
 #include "renderer_module.hpp"
