+++ engine/pathfinding/CMakeLists.txt
@@ -4,7 +4,6 @@ module_default_init(Pathfinding)
 target_link_libraries(Pathfinding
         PUBLIC Core
         PUBLIC Application
-        PUBLIC OldEngine
         PUBLIC Physics
         PUBLIC Renderer
 )
