+++ engine/renderer/CMakeLists.txt
@@ -18,3 +18,6 @@ target_link_libraries(Renderer
 target_include_directories(Renderer
         PUBLIC "${PROJECT_SOURCE_DIR}/external/stb"
 )
+
+source_exclude_unity("private/lib/stbi_impl.cpp")
+source_exclude_unity("private/lib/vma_impl.cpp")
