+++ engine/renderer/public/model_loader.hpp
@@ -38,5 +38,6 @@ public:
     void ReadGeometrySize(std::string_view path, uint32_t& vertexBufferSize, uint32_t& indexBufferSize);
 
 private:
+    constexpr static auto DEFAULT_LOAD_FLAGS = fastgltf::Options::DecomposeNodeMatrices | fastgltf::Options::LoadExternalBuffers | fastgltf::Options::LoadExternalImages;
     fastgltf::Parser _parser;
 };
