+++ engine/user_interface/private/ui_module.cpp
@@ -3,7 +3,6 @@
 #include "canvas.hpp"
 #include "renderer.hpp"
 #include "renderer_module.hpp"
-#include "ui_menus.hpp"
 
 ModuleTickOrder UIModule::Init(Engine& engine)
 {
